nbohr1more: yet more int to long conversions for 3954
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/x64@6664 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/Game_network.cpp
@@ -800,7 +800,7 @@ void idGameLocal::ServerProcessReliableMessage( int clientNum, const idBitMsg &m
 			break;
 		}
 		case GAME_RELIABLE_MESSAGE_VCHAT: {
-			int index = msg.ReadInt();
+			int index = msg.ReadLong();
 			bool team = msg.ReadBits( 1 ) != 0;
 			mpGame.ProcessVoiceChat( clientNum, team, index );
 			break;
@@ -838,7 +838,7 @@ void idGameLocal::ServerProcessReliableMessage( int clientNum, const idBitMsg &m
 
 			event->spawnId = msg.ReadBits( 32 );
 			event->event = msg.ReadByte();
-			event->time = msg.ReadInt();
+			event->time = msg.ReadLong();
 
 			event->paramsSize = msg.ReadBits( idMath::BitsForInteger( MAX_EVENT_PARAM_SIZE ) );
 			if ( event->paramsSize ) {
