+++ game/MultiplayerGame.cpp
@@ -1040,7 +1040,7 @@ void idMultiplayerGame::NewState( gameState_t news, idPlayer *player ) {
 
 			outMsg.Init( msgBuf, sizeof( msgBuf ) );
 			outMsg.WriteByte( GAME_RELIABLE_MESSAGE_WARMUPTIME );
-			outMsg.WriteInt( warmupEndTime );
+			outMsg.WriteLong( warmupEndTime );
 			networkSystem->ServerSendReliableMessage( -1, outMsg );
 
 			break;
@@ -3268,7 +3268,7 @@ void idMultiplayerGame::VoiceChat( const idCmdArgs &args, bool team ) {
 
 	outMsg.Init( msgBuf, sizeof( msgBuf ) );
 	outMsg.WriteByte( GAME_RELIABLE_MESSAGE_VCHAT );
-	outMsg.WriteInt( index );
+	outMsg.WriteLong( index );
 	outMsg.WriteBits( team ? 1 : 0, 1 );
 	networkSystem->ClientSendReliableMessage( outMsg );
 }
@@ -3333,7 +3333,7 @@ void idMultiplayerGame::ServerWriteInitialReliableMessages( int clientNum ) {
 	outMsg.WriteByte( GAME_RELIABLE_MESSAGE_STARTSTATE );
 	// send the game state and start time
 	outMsg.WriteByte( gameState );
-	outMsg.WriteInt( matchStartedTime );
+	outMsg.WriteLong( matchStartedTime );
 	outMsg.WriteShort( startFragLimit );
 	// send the powerup states and the spectate states
 	for( i = 0; i < gameLocal.numClients; i++ ) {
@@ -3356,7 +3356,7 @@ void idMultiplayerGame::ServerWriteInitialReliableMessages( int clientNum ) {
 	if ( gameState == COUNTDOWN ) {
 		outMsg.BeginWriting();
 		outMsg.WriteByte( GAME_RELIABLE_MESSAGE_WARMUPTIME );
-		outMsg.WriteInt( warmupEndTime );
+		outMsg.WriteLong( warmupEndTime );
 		networkSystem->ServerSendReliableMessage( clientNum, outMsg );
 	}
 }
@@ -3371,7 +3371,7 @@ void idMultiplayerGame::ClientReadStartState( const idBitMsg &msg ) {
 
 	// read the state in preparation for reading snapshot updates
 	gameState = (idMultiplayerGame::gameState_t)msg.ReadByte();
-	matchStartedTime = msg.ReadInt( );
+	matchStartedTime = msg.ReadLong( );
 	startFragLimit = msg.ReadShort( );
 	while ( ( client = msg.ReadShort() ) != MAX_CLIENTS ) {
 		assert( gameLocal.entities[ client ] && gameLocal.entities[ client ]->IsType( idPlayer::Type ) );
@@ -3386,6 +3386,6 @@ idMultiplayerGame::ClientReadWarmupTime
 ================
 */
 void idMultiplayerGame::ClientReadWarmupTime( const idBitMsg &msg ) {
-	warmupEndTime = msg.ReadInt();
+	warmupEndTime = msg.ReadLong();
 }
